+++ src/main.js
@@ -1,7 +1,10 @@
 'use babel';
 
 import { CompositeDisposable } from 'atom';
+import { getGlobalNavigationIntentions } from './intentions';
 import { firstHandlebarsComponentInLine } from './editor-utilities';
+import { getIntentions } from './helpers';
+import { getPossibleDestinations } from './helpers/editor';
 import {
   openComponent,
   toggleBetweenTestAndTarget,
@@ -14,6 +17,8 @@ export default {
   subscriptions: null,
 
   activate(state) {
+    require('atom-package-deps').install('ember-shortcuts');
+
     // Events subscribed to in atom's system can be easily cleaned up with a CompositeDisposable
     this.subscriptions = new CompositeDisposable();
 
@@ -46,6 +51,23 @@ export default {
 
   toggleControllerRoute() {
     toggleBetweenControllerAndRoute();
+  },
+
+  provideIntentions() {
+    return [
+      {
+        grammarScopes: ['text.html.handlebars'],
+        getIntentions: ({ textEditor, bufferPosition }) => {
+          return getIntentions(textEditor, bufferPosition);
+        }
+      },
+      {
+        grammarScopes: ['*'],
+        getIntentions: ({ textEditor, bufferPosition }) => {
+          return getGlobalNavigationIntentions(textEditor, bufferPosition);
+        }
+      }
+    ];
   }
 
 };