+++ src/strategies/abstractions/navigation.js
@@ -1,0 +1,16 @@
+'use babel';
+
+const fs = require('fs');
+
+class NavigationStrategy {
+  constructor(location) {
+    this.destinations = [];
+    this.location = location;
+  }
+
+  execute() {
+    return this.destinations.filter(destination => fs.existsSync(destination.path));
+  }
+}
+
+export default NavigationStrategy;